ID3v1 reading support
ID3v1 tags can now be read. Since the structure of ID3v1 tags is so
different from ID3v2 tags, the tags will be converted and stored
internally as ID3v2.3 tags. When saving the tag the ID3v1 tag will be
removed and the ID3v2 tag will be added.

Thanks to Olivier Renaud for the initial ID3v1 reading code!