Improve text rendering of Frame and Content
Add `name()` method to `Frame` which provides _human readable_
versions of the `Frame` Ids, e.g. `"TCOM" => "Composer"`.

Modify the `fmt::Display` implementation on `Frame` to return a string
of the form `"NAME = CONTENT"`.

Implement `fmt::Display` on `Content` and the struct and enums it
encapsulates.

Add `format_table()` method on `SynchronisedLyrics` which outputs a
plain text table.