Merge AIFF and WAV implementations
Both formats are chunk based, and use a ID3 chunk for tags. Minor differences include
endianness and chunk tags. The AIFF implementation did not handle padding, which caused it
to not work for some files. Beyond, the code used a temporary file when writing tags,
which is not optimal. The new generic implementation addresses all these issues, while
removing code duplication.