Merge AIFF and WAV implementations (#66)
* Merge AIFF and WAV implementations

Both formats are chunk based, and use a ID3 chunk for tags. Minor differences include
endianness and chunk tags. The AIFF implementation did not handle padding, which caused it
to not work for some files. Beyond, the code used a temporary file when writing tags,
which is not optimal. The new generic implementation addresses all these issues, while
removing code duplication.

* Add padding test

Chunk based files may include padding, which is not accounted for in chunk sizes. The new
AIFF implementation handles this detail, and the implemented test makes sure it works.