feat: Allow disabling expensive picture decoding
When decoding files, picture decoding is by far the most expensive part.
When running with mack[0], picture decoding alone accounts for 20% of
program runtime even though the pictures are never used. No other field
comes within even two orders of magnitude in terms of performance hit.

This commit adds a "decode_picture" feature, enabled by default, which
allows turning off this functionality if it is not needed in order to
allow increasing performance.

0: https://github.com/cdown/mack