feat: make file-related APIs compatible with io::Cursor
The codebase defines a StorageFile trait to model anything similar to a file.
By using `impl StorageFile` instead of `fs::File` in the API signatures, users
are able to provide either a file or something similar (for example, `io::Cursor`
to manipulate a block of data in memory).

This commit updates the APIs still using `fs::File` to `impl StorageFile`.